move loader files to their own directory
also, create specific directories for Windows and Linux
