+++ loader/windows/OpenCL.rc
