+++ loader/windows/icd_windows_hkr.h
